transformed auth reducers,action and type into one slice
