--- src/store/actions/auth.action.js
@@ -1,69 +1,0 @@
-import { URL_AUTH_SIGN_IN, URL_AUTH_SIGN_UP } from "../../constants/firebase";
-import { authTypes } from "../types";
-
-const {SIGN_UP, SIGN_IN} = authTypes;
-
-export const signUp = (email, password) => {
-    return async (dispatch) => {
-      try {
-        const response = await fetch(URL_AUTH_SIGN_UP, {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            email,
-            password,
-            returnSecureToken: true,
-            
-          }),
-        });
-  
-        if (!response.ok) {
-          throw new Error('Something went wrong!');
-        }
-  
-        const data = await response.json();
-  
-        dispatch({
-          type: SIGN_UP,
-          token: data.idToken,
-          userId: data.localId,
-          email: data.email,
-          
-        });
-      } catch (error) {
-        throw error;
-      }
-    };
-  };
-
-  export const signIn = (email, password) => {
-    return async (dispatch) => {
-      try {
-        const response = await fetch(URL_AUTH_SIGN_IN, {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify({
-            email,
-            password,
-            returnSecureToken: true,
-          }),
-        });
-  
-        const data = await response.json();
-        dispatch({
-          type: SIGN_IN,
-          token: data.idToken,
-          userId: data.localId,
-          email: data.email,
-          
-        });
-      } catch (error) {
-        throw error;
-      }
-    };
-  };
-
