--- src/store/reducers/index.js
@@ -1,2 +1,0 @@
-
-export { default as authReducer } from './auth.reducer';
