--- src/store/types/auth.types.js
@@ -1,4 +1,0 @@
-export const authTypes = {
-    SIGN_UP: 'SIGN_UP',
-    SIGN_IN: 'SIGN_IN',
-}
