--- src/store/types/index.js
@@ -1,2 +1,0 @@
-
-export * from './auth.types';
